+++ sbo_selenium/tests/test_errors.py
@@ -1,7 +1,6 @@
 from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
 
-from nose.tools import assert_raises_regexp
 from selenium.common.exceptions import TimeoutException
 
 from sbo_selenium import SeleniumTestCase
@@ -17,10 +16,10 @@ class TestErrors(SeleniumTestCase):
         """ It should be possible to set a custom error message for a condition timeout """
         self.get(reverse('good_accessibility'))
         msg = 'Custom message'
-        assert_raises_regexp(TimeoutException, msg, self.wait_for_condition, 'return no_such_variable', msg)
+        self.assertRaisesRegexp(TimeoutException, msg, self.wait_for_condition, 'return no_such_variable', msg)
 
     def test_condition_timeout_default_message(self):
         """ A condition timeout should have a reasonable default message"""
         self.get(reverse('good_accessibility'))
         msg = '"return no_such_variable" never became true'
-        assert_raises_regexp(TimeoutException, msg, self.wait_for_condition, 'return no_such_variable')
+        self.assertRaisesRegexp(TimeoutException, msg, self.wait_for_condition, 'return no_such_variable')